Move e2e test cases from /app folder to /e2e folder
