+++ internal/app/mmlogic/mmlogic_service_test.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io"
 	"testing"
 
 	structpb "github.com/golang/protobuf/ptypes/struct"
@@ -28,8 +27,6 @@ import (
 	"google.golang.org/grpc/status"
 	"open-match.dev/open-match/internal/config"
 	"open-match.dev/open-match/internal/pb"
-	"open-match.dev/open-match/internal/rpc"
-	rpcTesting "open-match.dev/open-match/internal/rpc/testing"
 	"open-match.dev/open-match/internal/statestore"
 	statestoreTesting "open-match.dev/open-match/internal/statestore/testing"
 )
@@ -163,67 +160,6 @@ func generateTickets(manifest1, manifest2 propertyManifest) []*pb.Ticket {
 	}
 
 	return testTickets
-}
-
-// TODO: move below test cases to e2e
-func TestQueryTicketsEmptyRequest(t *testing.T) {
-	assert := assert.New(t)
-	tc := createMmlogicForTest(t)
-	defer tc.Close()
-
-	queryTicketsLoop(t, tc, &pb.QueryTicketsRequest{}, func(_ *pb.QueryTicketsResponse, err error) {
-		assert.Equal(codes.InvalidArgument, status.Convert(err).Code())
-	})
-}
-
-func TestQueryTicketsForEmptyDatabase(t *testing.T) {
-	assert := assert.New(t)
-	tc := createMmlogicForTest(t)
-	defer tc.Close()
-
-	queryTicketsLoop(t, tc,
-		&pb.QueryTicketsRequest{
-			Pool: &pb.Pool{
-				Filter: []*pb.Filter{{
-					Attribute: "ok",
-				}},
-			},
-		},
-		func(resp *pb.QueryTicketsResponse, err error) {
-			assert.NotNil(resp)
-		})
-}
-
-func queryTicketsLoop(t *testing.T, tc *rpcTesting.TestContext, req *pb.QueryTicketsRequest, handleResponse func(*pb.QueryTicketsResponse, error)) {
-	c := pb.NewMmLogicClient(tc.MustGRPC())
-	stream, err := c.QueryTickets(tc.Context(), req)
-	if err != nil {
-		t.Fatalf("error querying tickets, %v", err)
-	}
-	for {
-		resp, err := stream.Recv()
-		if err == io.EOF {
-			break
-		}
-		handleResponse(resp, err)
-		if err != nil {
-			return
-		}
-	}
-}
-
-func createMmlogicForTest(t *testing.T) *rpcTesting.TestContext {
-	var closerFunc func()
-	tc := rpcTesting.MustServe(t, func(p *rpc.ServerParams) {
-		cfg := viper.New()
-		closerFunc = statestoreTesting.New(t, cfg)
-		cfg.Set("storage.page.size", 10)
-
-		BindService(p, cfg)
-	})
-	// TODO: This is very ugly. Need a better story around closing resources.
-	tc.AddCloseFunc(closerFunc)
-	return tc
 }
 
 func TestGetPageSize(t *testing.T) {
