+++ test/e2e/minimatch/minimatch_test.go
@@ -15,15 +15,25 @@
 package minimatch
 
 import (
+	"context"
 	"io"
 	"math"
 	"testing"
 	"time"
 
 	structpb "github.com/golang/protobuf/ptypes/struct"
+	"github.com/spf13/viper"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+	"open-match.dev/open-match/internal/app/backend"
+	"open-match.dev/open-match/internal/app/frontend"
+	"open-match.dev/open-match/internal/app/mmlogic"
 	"open-match.dev/open-match/internal/pb"
+	"open-match.dev/open-match/internal/rpc"
 	rpcTesting "open-match.dev/open-match/internal/rpc/testing"
+	statestoreTesting "open-match.dev/open-match/internal/statestore/testing"
 )
 
 const (
@@ -229,4 +239,276 @@ func testFetchMatches(assert *require.Assertions, poolTickets map[string][]strin
 			assert.Equal(gotTickets, poolTickets[br.Match.Roster[0].Name])
 		}
 	}
+}
+
+// TODO: refactor the tests below this line
+func TestAssignTickets(t *testing.T) {
+	assert := assert.New(t)
+	tc := createBackendForTest(t)
+	defer tc.Close()
+
+	fe := pb.NewFrontendClient(tc.MustGRPC())
+	be := pb.NewBackendClient(tc.MustGRPC())
+
+	ctResp, err := fe.CreateTicket(tc.Context(), &pb.CreateTicketRequest{Ticket: &pb.Ticket{}})
+	assert.Nil(err)
+
+	var tt = []struct {
+		req  *pb.AssignTicketsRequest
+		resp *pb.AssignTicketsResponse
+		code codes.Code
+	}{
+		{
+			&pb.AssignTicketsRequest{},
+			nil,
+			codes.InvalidArgument,
+		},
+		{
+			&pb.AssignTicketsRequest{
+				TicketId: []string{"1"},
+			},
+			nil,
+			codes.InvalidArgument,
+		},
+		{
+			&pb.AssignTicketsRequest{
+				TicketId: []string{"2"},
+				Assignment: &pb.Assignment{
+					Connection: "localhost",
+				},
+			},
+			nil,
+			codes.NotFound,
+		},
+		{
+			&pb.AssignTicketsRequest{
+				TicketId: []string{ctResp.Ticket.Id},
+				Assignment: &pb.Assignment{
+					Connection: "localhost",
+				},
+			},
+			&pb.AssignTicketsResponse{},
+			codes.OK,
+		},
+	}
+
+	for _, test := range tt {
+		resp, err := be.AssignTickets(tc.Context(), test.req)
+		assert.Equal(test.resp, resp)
+		if err != nil {
+			assert.Equal(test.code, status.Convert(err).Code())
+		} else {
+			gtResp, err := fe.GetTicket(tc.Context(), &pb.GetTicketRequest{TicketId: ctResp.Ticket.Id})
+			assert.Nil(err)
+			assert.Equal(test.req.Assignment.Connection, gtResp.Assignment.Connection)
+		}
+	}
+}
+
+func TestFetchMatches(t *testing.T) {
+	assert := assert.New(t)
+	tc := createBackendForTest(t)
+	defer tc.Close()
+
+	be := pb.NewBackendClient(tc.MustGRPC())
+
+	var tt = []struct {
+		req  *pb.FetchMatchesRequest
+		resp *pb.FetchMatchesResponse
+		code codes.Code
+	}{
+		{
+			&pb.FetchMatchesRequest{},
+			nil,
+			codes.InvalidArgument,
+		},
+	}
+
+	for _, test := range tt {
+		fetchMatchesLoop(t, tc, be, test.req, func(_ *pb.FetchMatchesResponse, err error) {
+			assert.Equal(test.code, status.Convert(err).Code())
+		})
+	}
+}
+
+// TODO: Add FetchMatchesNormalTest when Hostname getter used to initialize mmf service is in
+// https://github.com/GoogleCloudPlatform/open-match/pull/473
+func fetchMatchesLoop(t *testing.T, tc *rpcTesting.TestContext, be pb.BackendClient, req *pb.FetchMatchesRequest, handleResponse func(*pb.FetchMatchesResponse, error)) {
+	stream, err := be.FetchMatches(tc.Context(), req)
+	if err != nil {
+		t.Fatalf("error querying tickets, %v", err)
+	}
+	for {
+		resp, err := stream.Recv()
+		if err == io.EOF {
+			break
+		}
+		handleResponse(resp, err)
+		if err != nil {
+			return
+		}
+	}
+}
+
+func createBackendForTest(t *testing.T) *rpcTesting.TestContext {
+	var closerFunc func()
+	tc := rpcTesting.MustServe(t, func(p *rpc.ServerParams) {
+		cfg := viper.New()
+		closerFunc = statestoreTesting.New(t, cfg)
+
+		cfg.Set("storage.page.size", 10)
+		backend.BindService(p, cfg)
+		frontend.BindService(p, cfg)
+		mmlogic.BindService(p, cfg)
+	})
+	// TODO: This is very ugly. Need a better story around closing resources.
+	tc.AddCloseFunc(closerFunc)
+	return tc
+}
+
+// validateTicket validates that the fetched ticket is identical to the expected ticket.
+func validateTicket(t *testing.T, got *pb.Ticket, want *pb.Ticket) {
+	assert.Equal(t, got.Id, want.Id)
+	assert.Equal(t, got.Properties.Fields["test-property"].GetNumberValue(), want.Properties.Fields["test-property"].GetNumberValue())
+	assert.Equal(t, got.Assignment.Connection, want.Assignment.Connection)
+	assert.Equal(t, got.Assignment.Properties, want.Assignment.Properties)
+	assert.Equal(t, got.Assignment.Error, want.Assignment.Error)
+}
+
+// validateDelete validates that the ticket is actually deleted from the state storage.
+// Given that delete is async, this method retries fetch every 100ms up to 5 seconds.
+func validateDelete(t *testing.T, fe pb.FrontendClient, id string) {
+	start := time.Now()
+	for {
+		if time.Since(start) > 5*time.Second {
+			break
+		}
+
+		// Attempt to fetch the ticket every 100ms
+		_, err := fe.GetTicket(context.Background(), &pb.GetTicketRequest{TicketId: id})
+		if err != nil {
+			// Only failure to fetch with NotFound should be considered as success.
+			assert.Equal(t, status.Code(err), codes.NotFound)
+			return
+		}
+
+		time.Sleep(100 * time.Millisecond)
+	}
+
+	assert.Failf(t, "ticket %v not deleted after 5 seconds", id)
+}
+
+// TestFrontendService tests creating, getting and deleting a ticket using Frontend service.
+func TestFrontendService(t *testing.T) {
+	assert := assert.New(t)
+
+	tc := createStore(t)
+	fe := pb.NewFrontendClient(tc.MustGRPC())
+	assert.NotNil(fe)
+
+	ticket := &pb.Ticket{
+		Properties: &structpb.Struct{
+			Fields: map[string]*structpb.Value{
+				"test-property": {Kind: &structpb.Value_NumberValue{NumberValue: 1}},
+			},
+		},
+		Assignment: &pb.Assignment{
+			Connection: "test-tbd",
+		},
+	}
+
+	// Create a ticket, validate that it got an id and set its id in the expected ticket.
+	resp, err := fe.CreateTicket(context.Background(), &pb.CreateTicketRequest{Ticket: ticket})
+	assert.NotNil(resp)
+	assert.Nil(err)
+	want := resp.Ticket
+	assert.NotNil(want)
+	assert.NotNil(want.Id)
+	ticket.Id = want.Id
+	validateTicket(t, resp.Ticket, ticket)
+
+	// Fetch the ticket and validate that it is identical to the expected ticket.
+	gotTicket, err := fe.GetTicket(context.Background(), &pb.GetTicketRequest{TicketId: ticket.Id})
+	assert.NotNil(gotTicket)
+	assert.Nil(err)
+	validateTicket(t, gotTicket, ticket)
+
+	// Delete the ticket and validate that it was actually deleted.
+	_, err = fe.DeleteTicket(context.Background(), &pb.DeleteTicketRequest{TicketId: ticket.Id})
+	assert.Nil(err)
+	validateDelete(t, fe, ticket.Id)
+}
+
+func createStore(t *testing.T) *rpcTesting.TestContext {
+	var closer func()
+	tc := rpcTesting.MustServe(t, func(p *rpc.ServerParams) {
+		cfg := viper.New()
+		cfg.Set("playerIndices", []string{"testindex1", "testindex2"})
+		closer = statestoreTesting.New(t, cfg)
+		err := frontend.BindService(p, cfg)
+		assert.Nil(t, err)
+	})
+	tc.AddCloseFunc(closer)
+	return tc
+}
+
+// TODO: move below test cases to e2e
+func TestQueryTicketsEmptyRequest(t *testing.T) {
+	assert := assert.New(t)
+	tc := createMmlogicForTest(t)
+	defer tc.Close()
+
+	queryTicketsLoop(t, tc, &pb.QueryTicketsRequest{}, func(_ *pb.QueryTicketsResponse, err error) {
+		assert.Equal(codes.InvalidArgument, status.Convert(err).Code())
+	})
+}
+
+func TestQueryTicketsForEmptyDatabase(t *testing.T) {
+	assert := assert.New(t)
+	tc := createMmlogicForTest(t)
+	defer tc.Close()
+
+	queryTicketsLoop(t, tc,
+		&pb.QueryTicketsRequest{
+			Pool: &pb.Pool{
+				Filter: []*pb.Filter{{
+					Attribute: "ok",
+				}},
+			},
+		},
+		func(resp *pb.QueryTicketsResponse, err error) {
+			assert.NotNil(resp)
+		})
+}
+
+func queryTicketsLoop(t *testing.T, tc *rpcTesting.TestContext, req *pb.QueryTicketsRequest, handleResponse func(*pb.QueryTicketsResponse, error)) {
+	c := pb.NewMmLogicClient(tc.MustGRPC())
+	stream, err := c.QueryTickets(tc.Context(), req)
+	if err != nil {
+		t.Fatalf("error querying tickets, %v", err)
+	}
+	for {
+		resp, err := stream.Recv()
+		if err == io.EOF {
+			break
+		}
+		handleResponse(resp, err)
+		if err != nil {
+			return
+		}
+	}
+}
+
+func createMmlogicForTest(t *testing.T) *rpcTesting.TestContext {
+	var closerFunc func()
+	tc := rpcTesting.MustServe(t, func(p *rpc.ServerParams) {
+		cfg := viper.New()
+		closerFunc = statestoreTesting.New(t, cfg)
+		cfg.Set("storage.page.size", 10)
+
+		mmlogic.BindService(p, cfg)
+	})
+	// TODO: This is very ugly. Need a better story around closing resources.
+	tc.AddCloseFunc(closerFunc)
+	return tc
 }
